+++ node/tests/api/outline/getDocOutline.test.ts
@@ -1,0 +1,51 @@
+/**
+ * Copyright (C) 2024 SiYuan Community
+ * 
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import {
+    describe,
+} from "vitest";
+
+import client from "~/tests/utils/client";
+import { testKernelAPI } from "~/tests/utils/test";
+import { SchemaJSON } from "~/tests/utils/schema";
+
+import getDocOutline from "@/types/kernel/api/outline/getDocOutline";
+
+const pathname = client.Client.api.outline.getDocOutline.pathname;
+
+describe(pathname, async () => {
+    const schema_payload = new SchemaJSON(SchemaJSON.resolvePayloadSchemaPath(pathname));
+    const schema_response = new SchemaJSON(SchemaJSON.resolveResponseSchemaPath(pathname));
+    await schema_payload.loadSchemaFile();
+    await schema_response.loadSchemaFile();
+    const validate_payload = schema_payload.constructValidateFuction();
+    const validate_response = schema_response.constructValidateFuction();
+
+    testKernelAPI<getDocOutline.IPayload, getDocOutline.IResponse>({
+        name: "main",
+        payload: {
+            data: {
+                id: "20200813163359-v04n73b", // 隐私政策和用户协议
+            },
+            validate: validate_payload,
+        },
+        request: (payload) => client.client.getDocOutline(payload!),
+        response: {
+            validate: validate_response,
+        },
+    });
+});